HPCC-20133 Upgrade ESDL transaction log for Royalty logging, etc
1. Pass royalty dataset to log content in order to support
Royalty logging;
2. the transaction ID created by the existing code always has '-'
or '-X' between transaction seed and sequence number. Now, the
characters are configurable;
3. Create new class CLogContentFilter. Move the content of
CESPServerLoggingAgent.filterLogContent() into CLogContentFilter
.filterLogContent() so that it can be shared with other log
agents, such as WsLogServiceESPAgent. Also move the following
methods from CESPServerLoggingAgent to CLogContentFilter:
readAllLogFilters(), readLogFilters, addLogContentBranch,
filterAndAddLogContentBranch, and filterLogContentTree;
4. Move the following items from espserviceloggingagent.hpp to
loggingagentbase.hpp so that they can be shared for other log
agents: enum ESPLogContentGroup, const espLogContentGroupNames[],
class CTransIDBuilder, and class CESPLogContentGroupFilters.
5. Move wslogserviceespagent.xsl from LN repo to here because it
is needed by esp_service_DynamicESDL.xsl. Add code into
esp_service_DynamicESDL.xsl to call the template in
wslogserviceespagent.xsl
6. Move the xsl code of the transactionId/seed from logging_agent.xsl
to new esp_log_transid.xsl so that it can be shared by other log
agents.
