+++ esp/esdllib/esdl_transformer2.cpp
@@ -1831,6 +1831,8 @@ void Esdl2Transformer::processHPCCResult(IEspContext &ctx, IEsdlDefMethod &mthde
                 xppToXmlString(*xpp, stag, logdata);
             else if (strnicmp(dataset, "LOG_", 4)==0)
                 xppToXmlString(*xpp, stag, logdata);
+            else if (strieq(dataset, "royaltyset"))
+                xppToXmlString(*xpp, stag, logdata);
             else
             {
                 WARNLOG("ESDL processing HPCC Result: Dataset ignored: %s", dataset);
